# controllers/producer.go
package controllers

import (
	"net/http"
	"strings"
	"time"

	"github.com/julienschmidt/httprouter"
	"github.com/newscred/webhook-broker/storage"
	"github.com/newscred/webhook-broker/storage/data"
)

const (
	producersPath          = "/producers"
	producerIDPathParamKey = "producerId"
	producerPath           = "/producer/:" + producerIDPathParamKey
)

// MsgStakeholder is the
type MsgStakeholder struct {
	ID        string
	Name      string
	Token     string
	ChangedAt time.Time
}

// GetLastUpdatedHTTPTimeString exposes the string rep of the last modified timestamp for the object
func (stakeholder *MsgStakeholder) GetLastUpdatedHTTPTimeString() string {
	return stakeholder.ChangedAt.Format(http.TimeFormat)
}

func getMessageStakeholder(id string, stakeholderModel *data.MessageStakeholder) *MsgStakeholder {
	return &MsgStakeholder{ID: id, Name: stakeholderModel.Name, Token: stakeholderModel.Token, ChangedAt: stakeholderModel.UpdatedAt}
}

// ListResult is the resource returned by /producers endpoint
type ListResult struct {
	Result []string
	Pages  map[string]string
	Links  map[string]string
}

func isConditionalUpdateCalled(w http.ResponseWriter, r *http.Request, model data.Updateable) bool {
	validRequest := true
	unmodifiedSince := r.Header.Get(headerUnmodifiedSince)
	expectedHeader := model.GetLastUpdatedHTTPTimeString()
	if len(unmodifiedSince) <= 0 {
		writeBadRequest(w)
		validRequest = false
	} else if unmodifiedSince != expectedHeader {
		writePreconditionFailed(w)
		validRequest = false
	}
	return validRequest
}

func writeGetResult(err error, errFn func(w http.ResponseWriter), w http.ResponseWriter, msg interface{}) {
	if err != nil {
		errFn(w)
		return
	}
	if stakeholder, ok := msg.(data.Updateable); ok {
		w.Header().Add(headerLastModified, stakeholder.GetLastUpdatedHTTPTimeString())
	}
	writeJSON(w, msg)
}

func getUpdateData(r *http.Request, defaultName string) (token string, name string) {
	token = r.PostFormValue("token")
	if len(token) < 1 {
		token = randomToken()
	}
	name = r.PostFormValue("name")
	if len(name) < 1 {
		name = defaultName
	}
	return token, name
}

// ProducerController is for /producer/:prodId
type ProducerController struct {
	ProducerRepo storage.ProducerRepository
}

// Get implements the /producer/:prodId GET endpoint
// @Summary Get Producer Details
// @Description Retrieves details for a specific producer.
// @Tags Producers
// @Produce json
// @Param producerId path string true "Producer ID"
// @Success 200 {object} MsgStakeholder
// @Failure 404
// @Failure 500
// @Router /producer/{producerId} [get]
func (prodController *ProducerController) Get(w http.ResponseWriter, r *http.Request, param httprouter.Params) {
	producerID := param.ByName(producerIDPathParamKey)
	producerModel, err := prodController.ProducerRepo.Get(producerID)
	writeGetResult(err, writeNotFound, w, getMessageStakeholder(producerID, &producerModel.MessageStakeholder))
}

// Put implements the /producer/:prodId PUT endpoint
// @Summary Update Producer
// @Description Updates a producer's name and/or token.
// @Tags Producers
// @Accept x-www-form-urlencoded
// @Produce json
// @Param producerId path string true "Producer ID"
// @Param If-Unmodified-Since header string false "Only update if the resource has not been modified since this timestamp"
// @Param token formData string false "Token"
// @Param name formData string false "Name"
// @Success 200 {object} MsgStakeholder
// @Failure 400
// @Failure 412
// @Failure 415
// @Failure 500
// @Router /producer/{producerId} [put]
func (prodController *ProducerController) Put(w http.ResponseWriter, r *http.Request, param httprouter.Params) {
	validRequest := checkFormContentType(r, w)
	producerID := param.ByName(producerIDPathParamKey)
	producerModel, err := prodController.ProducerRepo.Get(producerID)
	if err == nil && validRequest {
		validRequest = isConditionalUpdateCalled(w, r, producerModel)
	}
	if !validRequest {
		return
	}
	token, name := getUpdateData(r, producerID)
	producer, _ := data.NewProducer(producerID, token)
	producer.Name = name
	producer, err = prodController.ProducerRepo.Store(producer)
	writeGetResult(err, func(w http.ResponseWriter) { writeErr(w, err) }, w, getMessageStakeholder(producerID, &producer.MessageStakeholder))
}

func checkFormContentType(r *http.Request, w http.ResponseWriter) bool {
	validRequest := true
	contentType, _, _ := strings.Cut(r.Header.Get(headerContentType), ";")
	if contentType != formDataContentTypeHeaderValue {
		validRequest = false
		writeUnsupportedMediaType(w)
	}
	return validRequest
}

// GetPath returns the endpoint's path
func (prodController *ProducerController) GetPath() string {
	return producerPath
}

// FormatAsRelativeLink Format as relative URL of this resource based on the params
func (prodController *ProducerController) FormatAsRelativeLink(params ...httprouter.Param) string {
	return formatURL(params, producerPath, producerIDPathParamKey)
}

// ProducersController for handling `/producers` endpoint
type ProducersController struct {
	ProducerRepo     storage.ProducerRepository
	ProducerEndpoint EndpointController
}

// Get implements the /producers endpoint
// @Summary List All Producers
// @Description Retrieves a paginated list of all producers.
// @Tags Producers
// @Produce json
// @Success 200 {object} ListResult
// @Failure 500
// @Router /producers [get]
func (prodController *ProducersController) Get(w http.ResponseWriter, r *http.Request, _ httprouter.Params) {
	producers, resultPagination, err := prodController.ProducerRepo.GetList(getPagination(r))
	if err != nil {
		writeErr(w, err)
		return
	}
	producerURLs := make([]string, len(producers))
	for index, producer := range producers {
		producerURLs[index] = prodController.ProducerEndpoint.FormatAsRelativeLink(httprouter.Param{Key: producerIDPathParamKey, Value: producer.ProducerID})
	}
	data := ListResult{Result: producerURLs, Pages: getPaginationLinks(r, resultPagination)}
	writeJSON(w, data)
}

// GetPath returns the endpoint's path
func (prodController *ProducersController) GetPath() string {
	return producersPath
}

// FormatAsRelativeLink Format as relative URL of this resource based on the params
func (prodController *ProducersController) FormatAsRelativeLink(params ...httprouter.Param) string {
	return producersPath
}

// NewProducerController initialize new producers controller
func NewProducerController(producerRepo storage.ProducerRepository) *ProducerController {
	return &ProducerController{ProducerRepo: producerRepo}
}

// NewProducersController initialize new producers controller
func NewProducersController(producerRepo storage.ProducerRepository, producerController *ProducerController) *ProducersController {
	return &ProducersController{ProducerRepo: producerRepo, ProducerEndpoint: producerController}
}