# docs/open-api-spec/swagger.yaml
definitions:
  config.SeedChannel:
    properties:
      id:
        description: ID is the ID of the data
        type: string
      name:
        description: Name is name of the data
        type: string
      token:
        description: Token is the pre configured token of the data
        type: string
    type: object
  config.SeedConsumer:
    properties:
      callbackURL:
        description: CallbackURL represents the URl to call back
        type: string
      channel:
        description: Channel represents which channel this consumer listens to
        type: string
      id:
        description: ID is the ID of the data
        type: string
      name:
        description: Name is name of the data
        type: string
      token:
        description: Token is the pre configured token of the data
        type: string
      type:
        type: string
    type: object
  config.SeedData:
    properties:
      channels:
        items:
          $ref: '#/definitions/config.SeedChannel'
        type: array
      consumers:
        items:
          $ref: '#/definitions/config.SeedConsumer'
        type: array
      dataHash:
        type: string
      producers:
        items:
          $ref: '#/definitions/config.SeedProducer'
        type: array
    type: object
  config.SeedProducer:
    properties:
      id:
        description: ID is the ID of the data
        type: string
      name:
        description: Name is name of the data
        type: string
      token:
        description: Token is the pre configured token of the data
        type: string
    type: object
  controllers.AppData:
    properties:
      appStatus:
        $ref: '#/definitions/data.AppStatus'
      seedData:
        $ref: '#/definitions/config.SeedData'
    type: object
...
  controllers.ListResult:
    properties:
      links:
        additionalProperties:
          type: string
        type: object
      pages:
        additionalProperties:
          type: string
        type: object
      result:
        items:
          type: string
        type: array
    type: object
...
  controllers.MsgStakeholder:
    properties:
      changedAt:
        type: string
      id:
        type: string
      name:
        type: string
      token:
        type: string
    type: object
...
  data.AppStatus:
    enum:
    - 1
    - 2
    - 3
    type: integer
    x-enum-varnames:
    - NotInitialized
    - Initializing
    - Initialized
...
info:
  contact: {}
  description: This is a fully HTTP-based Pub/Sub Broker designed to simplify system
    architecture in SOA or Microservice architecture.
  title: Webhook Broker API
  version: 1.0.0
paths:
  /_status:
    get:
      description: Retrieves the current status of the Webhook Broker application,
        including seed data and application state.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.AppData'
        "500":
          description: Internal Server Error
      summary: Get Application Status
      tags:
      - Status
...
  /job-status:
    get:
      description: Retrieves job status counts grouped by consumer.
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "500":
          description: Internal Server Error
      summary: Get Job Status Counts
      tags:
      - Status
  /producer/{producerId}:
    get:
      description: Retrieves details for a specific producer.
      parameters:
      - description: Producer ID
        in: path
        name: producerId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.MsgStakeholder'
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Get Producer Details
      tags:
      - Producers
    put:
      consumes:
      - application/x-www-form-urlencoded
      description: Updates a producer's name and/or token.
      parameters:
      - description: Producer ID
        in: path
        name: producerId
        required: true
        type: string
      - description: Only update if the resource has not been modified since this
          timestamp
        in: header
        name: If-Unmodified-Since
        type: string
      - description: Token
        in: formData
        name: token
        type: string
      - description: Name
        in: formData
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.MsgStakeholder'
        "400":
          description: Bad Request
        "412":
          description: Precondition Failed
        "415":
          description: Unsupported Media Type
        "500":
          description: Internal Server Error
      summary: Update Producer
      tags:
      - Producers
  /producers:
    get:
      description: Retrieves a paginated list of all producers.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.ListResult'
        "500":
          description: Internal Server Error
      summary: List All Producers
      tags:
      - Producers
securityDefinitions:
  ChannelToken:
    description: Channel authentication token issued when the channel was created
      or last rotated.
    in: header
    name: X-Broker-Channel-Token
    type: apiKey
  ConsumerToken:
    description: Consumer authentication token issued when the consumer was created
      or last rotated.
    in: header
    name: X-Broker-Consumer-Token
    type: apiKey
  ProducerID:
    description: Identifier of the producer originating the request.
    in: header
    name: X-Broker-Producer-ID
    type: apiKey
  ProducerToken:
    description: Producer authentication token issued when the producer was created
      or last rotated.
    in: header
    name: X-Broker-Producer-Token
    type: apiKey
swagger: "2.0"